+++ CHANGES.rst
@@ -18,8 +18,8 @@ Release History
    - Removed
    - Fixed
 
-0.6.0 (unreleased)
-------------------
+0.6.0 (December 13, 2017)
+-------------------------
 
 **Added**